+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [#1721](https://github.com/teambit/bit/issues/1721) enable removing/workspace-propagating a compiler/tester from component's config
 - [#1966](https://github.com/teambit/bit/issues/1966) prevent intermediate console printing when `--json` flag is used
 
 ## [14.2.6-dev.3] - 2019-09-03